usb: ehci: Add option to override ehci_readl()/ehci_writel()
There are EHCI-like hosts, like fotg210, out there that should
be able to use the shared ehci code but can't because they need
to override some parts of the ehci code to work properly.

Let ehci_readl() and ehci_writel() be overridden first.
This allows the ehci code to be used with hosts that need
to do special dances when accessing their registers.

The change is hidden behind a boolean kconfig option
so that the extra function pointer not null and call
only happens when a broken controller is enabled.